Move widgets for the Properties facet → `widgets.py`
